Fix UnicodeRanges. Improve anti-alias font render
The UnicodeRanges would use the wrong offset in the table for
adding code-points. The `std::lower_bound()` would select the entry
after the correct one. Now using `std::upper_bound()` and subtracting
one from the iterator.

The anti-aliasing in signed-distance-field rendering would use:
`distance / fwidth(distance)` for finding the distance from the edge.
`fwidth()` looks at the delta compared to neighbour's value. However
when the value doesn't change on line segments and in corners the delta
would be zero and thus wrong.

The new anti-alias algorithm uses a `distanceMultipler` passed via the
vertex buffer. And for each glyph the distanceMultipler is calculated
based on the font-size as displayed on the screen.