Rename MSDF to SDF since that is what is used.
The multichannel signed distance field was not a success for small text.
The lowest size of the texture-map of a glyph is determined by the
thinest feature on a font. In case of Time New Roman that required the
drawing of 64x64 pixels. multichannel signed distance field would use 32
bits per pixel in comparisong to 8 bits for a signed distance field.

For small text the 4 times increase in memory size does not weigh
against better rendering quality.