Fix text shaping, Add clipping, Start with shadow
Text shaping was bugged to how the lines widths where calculated.
Now do multiple steps, first create multiple lines and fold long lines.
Then calculated the width and height of each line.
Also properly align to the given box.

Adding clipping to the shader so that text can be clipped properly at
edges of widgets.

Start with working on shadow for fonts. First make the calculations of
the quads to match the atlas coordinates better. And make the border
around each atlas image based on the max distance of a SDF8 pixel.
Now the maximum shadow is based on the max-distance of a SDF8 pixel.