Remove pipeline dependence on window framebuffers.
Originally the pipeline would create resources for each frame buffer
of the window's swapchain's framebuffers; such as command buffers and
vertex buffers.

Now on each `render()` call the pipeline would directly build the
command buffer for the given framebuffer object. This allows the
pipeline to be indendend of the window.