Simplify widget render loop
Vertex data is now using the `vspan` data type which is simular to a
`std::span` but it can be used as a `vector` with the `push_back()`
`emplace_back()` and `clear()` methods. This way we do not need in pass
the offset seperately anymore.

The Pipeline-delegates have been removed and the Pipelines do not call
the delegates anymore to place the vertices. Instead the Window will
call `update()` with the vertex buffers of all pipelines.

Now each Widget can do all the drawing in the `update()` method.