Changes how modifier keys are handled.
Since we will have a translation between keys and text edit commands
I have made it so that modifiers are always linked to their relatives
names. In other words: Control is the control key on all computers,
Alt is the alt-, option- or meta-key. Super is the windows-, command- or
super-key. Beyond super is hyper, but most standard keyboards have a key
like that.