Implement keyboard-bindings parser
A simple JSON file parser which loads the key-bindings.
The key name is converted by the `KeyboardKey()` class;
the command is converted using tt5-encoding to a `string_ltag`
and the context is converted using tt5-encoding to a `string_tag`.

The `KeyboardEvent` now has a `getCommand()` function to retrieve the
command for the specific context that the specific widget is in.