Replace GLM in Image pipeline.
Updated the image-pipeline shader to use the same techniques as the
other shaders and use the new vec classes. Also change how Mouse
handling, window handling and box-model handling are done using the vec
classes.

In the future I want to change the rect classes to use two points,
allowing a cleaner implementation and 3D hit-box tests. Probably with
the same or better performance.