Rework of handleMouseEvent() and hitBoxTest()
The hitBoxTest() used to be only used by the operating system to check
where the mouse is inside the application, for example the system-menu
or resize corners. The new version is also used to detect which widget
is underneath the mouse cursor to send mouse events to. It is now a lot
more expensive since it needs to check if widgets like combo-box overlap
other widgets in a window, but it now also will change the mouse cursor
during the operating system hittest.

Because of widget overlap check I've implemented widget depth, starting
at 1.0 and subtracting 0.001 for each layer. A combo box when it starts
overlapping should subtract another 0.25.

The depth implementation caused a bug to show on the SDF-pipeline
drawing vertices in the wrong depth due to z-scaling with the font-size.
This was caused by vector/scaler multiplication. Changed scaler->vec
conversion to be explicit, and not alowing scaler mat::S().