Change keybinding system without context
Remove contexts from keybindings. Now all commands from a certain
keypress are send to a widget's `handleCommand()` method.
The widget can than choose which command it will handle.

Should the keyboard event simply hold all commands which the widgets
then modify and pop from the command list when handled? This requires
the keyboard event to be passed as non-const.