ShapedText is now for a single paragraph
We want ShapedText to handle one paragraph at a time, this means that
the calculations only have to be done on a current paragraph as this is
the unit where lines may get split based on the width of the text.

GlyphMetrics will now be scaled to the font-size of the glyph by the
shaped text. This means that instead of a transform the attributed-glyph
will now have a position instead, a position within the paragraph.

Widgets will still need to vertically and horizontally align the shaped
text within the widget.