EditableText handles selection and copy/cut/paste
For pasting to work, I needed to convert UTF-16 strings to graphemes
due to templated stringTranslate() functions this was fragile. Replaced
all UTF-8, UTF-16, UTF-32 and grapheme translations with to_xstring()
functions which are more stable.

Added selection in EditableText by keeping a cursor- and selection-index
when both are equal no text is selected, otherwise text is selected.
On movement selection is reset and on change selections are deleted.