Make LineInput widget cursor blink.
To make the widget cursor blink, I've added the displayClock to the
vertical-sync class and to the window and widget updateVertices()
functions. To add a timer event for rendering the window, I've created a
Trigger class that widgets and windows use to check if they need to be
rendered right now, or at a later time, and how much of the widget needs
to be recalculated.

Due to the blinking I found that the clock was not calibrated quickly
enough which showed as a fast blinking cursor until calibration. I make
a 100ms calibration at the start.

Due to adding the displayClock to the vertical-sync class I found that
the clock conversion was slow, I updated the bigint class to be more
modern and made some optimization, such as force inlining some
functions, it helped a lot; however not nearly enough. So I changed the
clock conversion algorithm to not use bigint anymore and build a
specific implementation using wide_multiply() and two shifts.