Optimize matrix transforms.
Created a mat::T translation and mat::S scale types. These types can be
used as if they are standard mat types, but are more efficient in
storage and computational with doing matrix/matrix and matrix/vector
multiplications.

Included is a is_mat<> type trait, which can be used with std::enable_if<>.

The Path, BezierPoint, BezierCurve and GlyphMetrics have been optimized
to use the is_mat<> as well.