Widget drawing is now done through DrawContext.
The new DrawContext class is used to add FilledQuads, Boxes (with border
and shaped corners), Text and Images to the Vertex buffers of the
shaders.

The DrawContext will keep track of color, fillColor, borderColor,
borderSize, shadowSize, clippingRectangle, and transformations.
It is expected that the widget makes a local copies of DrawContext so that
it may modify settings locallly.