Replaced library singletons with startup/shutdown
Removed the original singletons in the libraries (TTauri, TTauri::Text,
TTauri::GUI, TTauri::GUI::Widgets and TTauri::Audio) with startup() and
shutdown() functions for each library. Each startup() will call the
startup of the library it will need to function, for this each startup()
shutdown() pair will have a reference counter to only execute once.

All the variables inside the singletons have been promoted to actual
global variables.

The main objects for the audio and gui systems are now called
`audioSystem` and `guiSystem`.