Add theming.
The GUI can now be themed using the Theme class. At startup of the
Application I load all the theme.json files in a Theme instance inside
the ThemeBook. In globals the ThemeBook is initialized to select the
current dark/light mode from the operating system and when the
dark/light mode changes a theme is searched to match the theme name and
theme mode as closely as possible.

We have not implemented a way for the user to select a theme yet.