Make toggle switch, with value observing.
Create a value observing classes, and let a toggle widget observe a
boolean value. On top of this I am using an animation class to
handle animations of observed values.

Currently mixing the color of the button to fade in and out on a toggle,
but in the future I will create an toggle button graphic for this.