Add checkbox, radio buttons and toggle switch
Completed work on checkbox, radio buttons and toggle switch widgets.
The checkbox and radio button will work different numeric types/enum.
As long as they can be passed as non-type template arguments.

Also changed rendering to pre-multiplied alpha. And found a bug in the
text rendering which was doing pre-multiplication twice, which made it
look better. Can get it to work well enough, light on dark makes the
text to thick, as if vulkan isn't blending in linear colours, instead
taking the srgb pixel values as is.

Need to do some expirimentation with full black and white rendering and
renderdoc to diagnose.