Use perceptual compositing for text.
Even though alpha compositing should be done in linear space, the alpha
value itself should be calculated using perceptional transfer function
based on the brighness of the two colors that are being composited.

The above is required so that the thickness of the anti-aliasing feels
equal for both dark-on-light and light-on-dark text. It is very hard to
do this correctly, so instead we just look at the text color and expect
it to contrast with the background. And use linear, gamma 2.0 or
inv-gamma 2.0 alpha value depending on the brightness of the text
(the green component).

We also need to compensate the glyph thickness slighly for the now
correct rendering of the glyph compared to design of the font.