Fix tone mapper bug.
The tone mapper had a bug where the descriptor set, which included a
reference to the input attachment's image view, was not updated when the
swap chain was rebuild. This swap chain rebuild causes the input
attachment to resize as well, and therefor the descriptor set was no
longer pointing to a valid image view.

I've updated this so that all pipelines will not rebuild the descriptor
set when the pipeline gets rebuild do to swapchain change.

To debug this I have also included the RenderDoc API to inject during
startup with debug builds. This allows the ability to debug the
application when it crashes while running inside RenderDoc.