Improve SDF subpixel rendering.
Fixed an issue with the wrong font-variant being chosen, for example
Times New Roman would select the bold version when a regular was
requested.

The subpixel math is still based on SDF distance being purely
responsable for how much the full-pixel is covered. This means that
pixels are now sub-pixel positioned they are still a full pixel wide.

The effect is pretty subtle, simular to strong sub-pixel rendering with
a post processing low-pass filter.