Remove dependence of font size in SDF shader
Instead of using the font size I am using the size of the atlas and the
fwidth() of the texture coordinates to figure out by how much to scale
the distance sampled from the atlas. By removing this dependence on the
font size scaling and rotation with transformation matrix should work
more correctly, and also open the ability to do seperate font scaling.

Also make it possible to select LCD subpixel orientation at the window
level.