Observer works without observed.
The observer objects now work without an observed object,
meaning it can use its internal value instead. This means
we can use the observer on any widget attribute that the widget needs to
monitor, without having to link it to a value outside of the widget.

Also start work on rounding boxes to pixel boundaries. Add documentation
about DPI scaling.