Change coordinate system to be more consistant.
The local coordinate system of a window and widget now has their origin
in the far left-bottom corner, including the left-bottom corner of the
point and pixel.

We also defined DPI scaling to be in integer values, and made it the
responsibility of the widget to draw borders through the center of the
pixels. The DrawContext includes a helper function to draw a box,
including the border, within the given corner coordinates to help with
drawing the border sharply.