Add observed values
Attributes of widgets will be observers which can be linked to
observable values. When a observable value is modified the widget is
notified and redrawn.

I've also included an expression language in which you can make
expressions out of observable value and link these to the widgets. Some
of the expressions are reverseable, so that for example a toggle button
can function completely inverse of a observed boolean through a
not-expression.