Improve text input widget.
Fixed bug so that the mouse select position is where the text is located.

Scroll text that is too large to fit in the widget. This required the
implementation of non-wrap in the text shaper while still giving the
text shaper the actual size of the widget to properly align the text
that is now to large.

Implement tripple click paragraph selection and drag.