Implement mouse dragging
Widgets will now receive a seperate drag events, which used to be
move events. drag events will be send to the widget where the drag
started even when the drag continues outside of the widget. button-up
events after the drag will also be send to the widget where the drag
started.

Updated all button like widgets to check if on button-up if the drag
terminated inside the widget before activating.

Update the line input widget to handle drags outside of the widget and
do an animated scroll and select characters/words/paragraphs at the
edge of the widget.