Start work on SystemMenuWidget.
The system menu widget will show the application icon in the place of a
Windows 10 system-menu-button.

This widget will use a PixelMap to show an image. This required the
PipelineImage to be rewritten to no longer use a cach and no longer
handling BackingImages. Instead a much simpler system of allocation
of images and uploading of images to the atlas has been created.

At the same time I modified the PipelineImage system to work with
R16G16B16A16SFloat images.