Fix #522 get_keyboard_modifiers narrow_cast fails
Not sure why it wasn't detected before, because pressing the shift
key will crash any HikoGUI application. I know I have tested
shift-tab before when doing keyboard-navigation.
The get_keyboard_modifiers hasn't changes in many years. So the
only thing I can think is that narrow_cast is now more stricter.

In any case, after reading the documentation for GetAsyncKeyState()
I deterimined it was best to check for negative result, as it is
the most significant bit of the result that is used to determine
if the key was pressed. It looks like this function was actually
designed to be used like this.

Also added a static_assert to make sure that GetAsyncKeyState()
actually returns a signed-integer.