vk::acquireNextImageKHR() eNotReady bug (#532)
* Possible fix #529

vk::acquireNextImageKHR() can return eNotReady. This is a success
value, meaning that the image was acquired; the image is just not
ready yet.

Before the image can be used we need to wait on the semaphore passed
to vk::acquireNextImageKHR(). This should already be handled
correctly by HikoGUI.

Since we can not see this behavior on the current computer this is
only a potential fix.