fix selection box (#638)
This patch fixes two bugs.

Scrolling the selection box caused the selection box to disappear.
The event for button-up was not handled by the scroll-bar and instead was chained to the ancestor the selection button which closed the selection box in response.

This was solved by shortcutting any events in the overlay widget, so that no events would go deeper.

Using the scroll wheel on the selection box menu buttons would cause the wrong view to scroll.
The buttons where parented on the selection-widget instead of the grid-widget that was holding those buttons.

Changed the selection-delegate to accept a parent for the widgets, so that the grid-widget now can become the parent. This solved a lot of weird issues with the selection buttons.